% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{distinct_colors}
\alias{distinct_colors}
\title{Distinct colors for large categorical datasets}
\usage{
distinct_colors(n)
}
\arguments{
\item{n}{Number of colors desired}
}
\value{
Color palette
}
\description{
Colors derived from the R package colorspace
colorspace: A Toolbox for Manipulating and Assessing Colors and Palettes
https://cran.r-project.org/web/packages/colorspace/index.html
License: BSD_3_clause + https://cran.r-project.org/web/packages/colorspace/LICENSE
}
