% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{do_rwr}
\alias{do_rwr}
\title{Network propagation wrapper}
\usage{
do_rwr(ig, seeds, restart = 0.5, normalise = "laplacian", verbose = FALSE, ...)
}
\arguments{
\item{ig}{An igraph object}

\item{seeds}{One or more seed nodes}

\item{restart}{Restart probability between 0-1}

\item{normalise}{Normalization method for the adjacency matrix}

\item{verbose}{Use false for quiet mode}

\item{...}{Additional keyword arguments}
}
\value{
Propagation values
}
\description{
Network propagation wrapper
}
