% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{do_signature}
\alias{do_signature}
\title{Extend a signature in the context of a network}
\usage{
do_signature(ig, signature, restart = 0.5)
}
\arguments{
\item{ig}{An igraph object}

\item{signature}{A character vector of symbols}

\item{restart}{Restart probability between 0-1}
}
\value{
An igraph and an extended signature
}
\description{
Extend a signature in the context of a network
}
