% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich_communities}
\alias{enrich_communities}
\title{Do enrichment for all communities of the network}
\usage{
enrich_communities(ig, genesets, restart = 0.5, fdr = 1, pval = 1)
}
\arguments{
\item{ig}{An igraph object}

\item{genesets}{A named list of genesets}

\item{restart}{Restart probability between 0-1}

\item{fdr}{Passed to hypeR}

\item{pval}{Passed to hypeR}
}
\value{
A list of hyp objects
}
\description{
Do enrichment for all communities of the network
}
