% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich_communities_static}
\alias{enrich_communities_static}
\title{Static enrichment for all communities of the network}
\usage{
enrich_communities_static(
  ig,
  genesets,
  restart = 0.5,
  fdr = 1,
  pval = 1,
  top = 1,
  val = NULL
)
}
\arguments{
\item{ig}{An igraph object}

\item{genesets}{A named list of genesets}

\item{restart}{Restart probability between 0-1}

\item{fdr}{Passed to hypeR}

\item{pval}{Passed to hypeR}

\item{top}{Limit number of genesets shown}

\item{val}{A value to associate with genesets}
}
\value{
An igraph object
}
\description{
Static enrichment for all communities of the network
}
