% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{extend_signature}
\alias{extend_signature}
\title{Extend a signature in the context of a network}
\usage{
extend_signature(ig, signature, restart = 0.5)
}
\arguments{
\item{ig}{An igraph object}

\item{signature}{A character vector of node identifiers}

\item{restart}{Restart probability between 0-1}
}
\value{
An igrpah object
}
\description{
Extend a signature in the context of a network
}
