% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{labelize_communities}
\alias{labelize_communities}
\title{Labelize community enrichment}
\usage{
labelize_communities(communities.hyp, top = 1, sep = "\\n", val = NULL)
}
\arguments{
\item{communities.hyp}{A list of hyp objects}

\item{top}{Limit number of genesets shown}

\item{sep}{label delimeter}

\item{val}{A value to associate with genesets}
}
\value{
A list of community enrichment labels
}
\description{
Labelize community enrichment
}
