% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{distinct_colors}
\alias{distinct_colors}
\title{Distinct colors for large categorical datasets}
\usage{
distinct_colors(n)
}
\arguments{
\item{n}{Number of colors desired}
}
\value{
Color palette
}
\description{
Distinct colors for large categorical datasets
}
